fix(tests): Move directory `tests\Providers` to `tests\Support\Provider` for better organization.
